// frontend/src/app/core/api/model/promotion-subject-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PromotionSubjectResponse { 
    id?: number;
    subjectName?: string;
    theoryHours?: number;
    practicalHours?: number;
    totalHours?: number;
    coefficient?: number;
}