// frontend/src/app/core/api/model/assessment-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AssessmentResponse { 
    id?: number;
    promotionSubjectId?: number;
    subjectName?: string;
    title?: string;
    assessmentType?: AssessmentResponse.AssessmentTypeEnum;
    totalMarks?: number;
    weightPercentage?: number;
    status?: AssessmentResponse.StatusEnum;
}
export namespace AssessmentResponse {
    export const AssessmentTypeEnum = {
        Ds: 'DS',
        Tp: 'TP',
        FinalExam: 'FINAL_EXAM',
        Retake: 'RETAKE'
    } as const;
    export type AssessmentTypeEnum = typeof AssessmentTypeEnum[keyof typeof AssessmentTypeEnum];
    export const StatusEnum = {
        Draft: 'DRAFT',
        Planned: 'PLANNED',
        GradingInProgress: 'GRADING_IN_PROGRESS',
        SubmittedToAdmin: 'SUBMITTED_TO_ADMIN',
        Published: 'PUBLISHED',
        Cancelled: 'CANCELLED',
        Locked: 'LOCKED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}